requirements.txt
django
djangorestframework
google-auth==2.36.0
requests
